90% of the work is done!

+++ frontend/delete.php
@@ -1,5 +1,12 @@
 <?php
-$id = reset($_GET);
+$id = ($_GET['id'] ?? '');
+$teamID = ($_GET['teamID'] ?? '');
 require_once(__DIR__ . "/../backend/utils/storage.inc.php");
 $commentsStorage = new Storage(new JsonIO(__DIR__ . "/../backend/data/comments.json"));
-$commentsStorage->delete($id);
+if($id != '' and $teamID != ''){
+    $commentsStorage->delete($id);
+    header("Location: ./team_details.php?teadID=".$teamID, true, 301);
+    exit();
+}else{
+    echo 'Link error!';
+}
